[w2llatexform][w2lonpdf] Style changes. CSS id and class added.
New CSS: '#w2lFmFormer' is the id of the form in the 'w2llatexform' page and
'.w2lLogOutput' is the class of the textareas in the 'onpdf' page.
There is no defined style for those elements, their CSS code must be written
in a skin of the wiki.

Added two messages:
- 'wiki2latex_version' which shows the extension name with its version.
- 'wiki2latex_version_tag' shows the extension name between the passed
parameters. The first one is meant to be the opening tag and the other the
closing tag. This allows to write (e.g.) "&lt;a href="site.com"&gt; Wiki2Latex
&lt;/a&gt; (W2L_VERSION)".